test (LottoServiceTest.java) : 수익률 계산 테스트
