+++ src/test/java/lotto/service/LottoServiceTest.java
@@ -1,13 +1,24 @@
 package lotto.service;
 
 import lotto.domain.Money;
+import lotto.domain.item.LottoNumber;
+import lotto.domain.item.LottoNumbers;
+import lotto.domain.item.LottoTickets;
+import lotto.domain.item.WinLottoTicket;
+import lotto.view.StatisticsLottoDto;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import java.util.ArrayList;
+import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoServiceTest {
+
+    LottoNumbers winNumber = LottoNumbers.createLottoNumbersUseInteger(new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 6)));
 
     @DisplayName("구매할 금액과 수동으로 사려는 로또의 수량 비교")
     @ParameterizedTest
@@ -21,5 +32,22 @@ class LottoServiceTest {
 
         //then
         assertThat(result).isEqualTo(Integer.parseInt(count));
+    }
+
+
+    @DisplayName("수익률 개산")
+    @Test
+    public void getEarningRate_success() throws Exception {
+        //given
+        LottoService service = new LottoService();
+        final WinLottoTicket winLottoTicket = new WinLottoTicket(winNumber, new LottoNumber(45));
+        final LottoTickets lottoToPassivity =
+                LottoTickets.createLottoToPassivity(Arrays.asList(LottoNumbers.of(Arrays.asList(1, 2, 3, 10, 11, 12))));
+
+        //when
+        StatisticsLottoDto earningRate = service.getEarningRate(lottoToPassivity, winLottoTicket);
+
+        //then
+        assertThat(earningRate.getEarningRate()).isEqualTo(5);
     }
 }
